Replace WebView URL with one that supports JS API
As the title says, the WebView URI used doesn't work with the JavaScript API mentioned in the LiveChat docs. Specifically, it doesn't expose access to LC_API, the global object with LiveChat's functions. Using the new link enables that access, so apps can inject JS to update user's name, email, and other variables, without filling in a redundant form.

Also I think it might come with performance improvements on older devices? Not 100% on that one but it feels faster in testing on my old Android.